Moved actions tests and fixed errors
