+++ app/components/MovieCard/MovieCard.test.js
@@ -1,0 +1,6 @@
+describe(`MovieCard Component Tests`, () => {
+
+  it(`should have a test`, () => {
+    expect(true).toBe(false);
+  });
+});
