+++ app/containers/LoginContainer/LoginContainer.js
@@ -1,0 +1,2 @@
+import { connect } from 'react-redux';
+import { tryLogin, createUser } from './actions';
