+++ app/tests/MovieCard.test.js
@@ -1,5 +1,5 @@
 describe('MOVIE CARD', () => {
   it.skip('should return false', () => {
     expect(false).toBe(true);
-  })
-})
+  });
+});
